Refactor: Move atomic operation classes to new package
All atomic operation classes have been moved from `jromp.parallel.construct.atomic.operation` to `jromp.parallel.operation` for better organization and clarity. Import statements in the corresponding files have been updated to reflect this change.
