+++ src/main/java/jromp/parallel/operation/Operation.java
@@ -1,4 +1,4 @@
-package jromp.parallel.construct.atomic.operation;
+package jromp.parallel.operation;
 
 import java.io.Serializable;
 import java.util.function.Supplier;
