+++ src/test/java/jromp/parallel/construct/atomic/operation/AllAtomicOperationsTests.java
@@ -2,6 +2,7 @@ package jromp.parallel.construct.atomic.operation;
 
 import jromp.parallel.Parallel;
 import jromp.parallel.construct.atomic.Atomic;
+import jromp.parallel.operation.Operations;
 import jromp.parallel.var.SharedVariable;
 import jromp.parallel.var.Variables;
 import org.junit.jupiter.api.Test;
